Targetting Update. Effect Tests
Targetting seems to work now. Had an epiphany that I was making it
harder on myself than it needed to be. Can only attack creatures that
have been cast, that arent yourself, or you cna attack a player.
friendly fire IS something you can do at this point in time. I kind of
like it. Allows for  some unique possibilities.
Added a few more test creatures with different effects. (Turn start, on
attack)